GRLDarknet inherits from Darknet, remove repeat code, add __init__ files for imports
