+++ src/models/yolo_model.py
@@ -1,4 +1,5 @@
 from __future__ import division
+import os
 
 import torch
 import torch.nn as nn
@@ -6,7 +7,7 @@ import torch.nn.functional as F
 from torch.autograd import Variable
 import numpy as np
 
-from util.util_yolo import build_targets, to_cpu, non_max_suppression
+from src.util.util_yolo import build_targets, to_cpu, non_max_suppression
 
 import matplotlib.pyplot as plt
 import matplotlib.patches as patches
@@ -17,10 +18,10 @@ def parse_model_config(path):
     file = open(path, 'r')
     lines = file.read().split('\n')
     lines = [x for x in lines if x and not x.startswith('#')]
-    lines = [x.rstrip().lstrip() for x in lines] # get rid of fringe whitespaces
+    lines = [x.rstrip().lstrip() for x in lines]  # get rid of fringe whitespaces
     module_defs = []
     for line in lines:
-        if line.startswith('['): # This marks the start of a new block
+        if line.startswith('['):  # This marks the start of a new block
             module_defs.append({})
             module_defs[-1]['type'] = line[1:-1].rstrip()
             if module_defs[-1]['type'] == 'convolutional':
@@ -48,74 +49,7 @@ def parse_data_config(path):
     return options
 
 
-def create_modules(module_defs):
-    """
-    Constructs module list of layer blocks from module configuration in module_defs
-    """
-    hyperparams = module_defs.pop(0)
-    output_filters = [int(hyperparams["channels"])]
-    module_list = nn.ModuleList()
-    for module_i, module_def in enumerate(module_defs):
-        modules = nn.Sequential()
 
-        if module_def["type"] == "convolutional":
-            bn = int(module_def["batch_normalize"])
-            filters = int(module_def["filters"])
-            kernel_size = int(module_def["size"])
-            pad = (kernel_size - 1) // 2
-            modules.add_module(
-                f"conv_{module_i}",
-                nn.Conv2d(
-                    in_channels=output_filters[-1],
-                    out_channels=filters,
-                    kernel_size=kernel_size,
-                    stride=int(module_def["stride"]),
-                    padding=pad,
-                    bias=not bn,
-                ),
-            )
-            if bn:
-                modules.add_module(f"batch_norm_{module_i}", nn.BatchNorm2d(filters, momentum=0.9, eps=1e-5))
-            if module_def["activation"] == "leaky":
-                modules.add_module(f"leaky_{module_i}", nn.LeakyReLU(0.1))
-
-        elif module_def["type"] == "maxpool":
-            kernel_size = int(module_def["size"])
-            stride = int(module_def["stride"])
-            if kernel_size == 2 and stride == 1:
-                modules.add_module(f"_debug_padding_{module_i}", nn.ZeroPad2d((0, 1, 0, 1)))
-            maxpool = nn.MaxPool2d(kernel_size=kernel_size, stride=stride, padding=int((kernel_size - 1) // 2))
-            modules.add_module(f"maxpool_{module_i}", maxpool)
-
-        elif module_def["type"] == "upsample":
-            upsample = Upsample(scale_factor=int(module_def["stride"]), mode="nearest")
-            modules.add_module(f"upsample_{module_i}", upsample)
-
-        elif module_def["type"] == "route":
-            layers = [int(x) for x in module_def["layers"].split(",")]
-            filters = sum([output_filters[1:][i] for i in layers])
-            modules.add_module(f"route_{module_i}", EmptyLayer())
-
-        elif module_def["type"] == "shortcut":
-            filters = output_filters[1:][int(module_def["from"])]
-            modules.add_module(f"shortcut_{module_i}", EmptyLayer())
-
-        elif module_def["type"] == "yolo":
-            anchor_idxs = [int(x) for x in module_def["mask"].split(",")]
-            # Extract anchors
-            anchors = [int(x) for x in module_def["anchors"].split(",")]
-            anchors = [(anchors[i], anchors[i + 1]) for i in range(0, len(anchors), 2)]
-            anchors = [anchors[i] for i in anchor_idxs]
-            num_classes = int(module_def["classes"])
-            img_size = int(hyperparams["height"])
-            # Define detection layer
-            yolo_layer = YOLOLayer(anchors, num_classes, img_size)
-            modules.add_module(f"yolo_{module_i}", yolo_layer)
-        # Register module list and number of output filters
-        module_list.append(modules)
-        output_filters.append(filters)
-
-    return hyperparams, module_list
 
 
 class Upsample(nn.Module):
@@ -272,7 +206,7 @@ class Darknet(nn.Module):
     def __init__(self, config_path, img_size=416):
         super(Darknet, self).__init__()
         self.module_defs = parse_model_config(config_path)
-        self.hyperparams, self.module_list = create_modules(self.module_defs)
+        self.hyperparams, self.module_list = self.create_modules(self.module_defs)
         self.yolo_layers = [layer[0] for layer in self.module_list if hasattr(layer[0], "metrics")]
         self.img_size = img_size
         self.seen = 0
@@ -298,6 +232,76 @@ class Darknet(nn.Module):
         yolo_outputs = to_cpu(torch.cat(yolo_outputs, 1))
         return yolo_outputs if targets is None else (loss, yolo_outputs)
 
+    def create_modules(self,module_defs):
+        """
+        Constructs module list of layer blocks from module configuration in module_defs
+        """
+        hyperparams = module_defs.pop(0)
+        output_filters = [int(hyperparams["channels"])]
+        module_list = nn.ModuleList()
+        for module_i, module_def in enumerate(module_defs):
+            modules = nn.Sequential()
+
+            if module_def["type"] == "convolutional":
+                bn = int(module_def["batch_normalize"])
+                filters = int(module_def["filters"])
+                kernel_size = int(module_def["size"])
+                pad = (kernel_size - 1) // 2
+                modules.add_module(
+                    f"conv_{module_i}",
+                    nn.Conv2d(
+                        in_channels=output_filters[-1],
+                        out_channels=filters,
+                        kernel_size=kernel_size,
+                        stride=int(module_def["stride"]),
+                        padding=pad,
+                        bias=not bn,
+                    ),
+                )
+                if bn:
+                    modules.add_module(f"batch_norm_{module_i}", nn.BatchNorm2d(filters, momentum=0.9, eps=1e-5))
+                if module_def["activation"] == "leaky":
+                    modules.add_module(f"leaky_{module_i}", nn.LeakyReLU(0.1))
+
+            elif module_def["type"] == "maxpool":
+                kernel_size = int(module_def["size"])
+                stride = int(module_def["stride"])
+                if kernel_size == 2 and stride == 1:
+                    modules.add_module(f"_debug_padding_{module_i}", nn.ZeroPad2d((0, 1, 0, 1)))
+                maxpool = nn.MaxPool2d(kernel_size=kernel_size, stride=stride, padding=int((kernel_size - 1) // 2))
+                modules.add_module(f"maxpool_{module_i}", maxpool)
+
+            elif module_def["type"] == "upsample":
+                upsample = Upsample(scale_factor=int(module_def["stride"]), mode="nearest")
+                modules.add_module(f"upsample_{module_i}", upsample)
+
+            elif module_def["type"] == "route":
+                layers = [int(x) for x in module_def["layers"].split(",")]
+                filters = sum([output_filters[1:][i] for i in layers])
+                modules.add_module(f"route_{module_i}", EmptyLayer())
+
+            elif module_def["type"] == "shortcut":
+                filters = output_filters[1:][int(module_def["from"])]
+                modules.add_module(f"shortcut_{module_i}", EmptyLayer())
+
+            elif module_def["type"] == "yolo":
+                anchor_idxs = [int(x) for x in module_def["mask"].split(",")]
+                # Extract anchors
+                anchors = [int(x) for x in module_def["anchors"].split(",")]
+                anchors = [(anchors[i], anchors[i + 1]) for i in range(0, len(anchors), 2)]
+                anchors = [anchors[i] for i in anchor_idxs]
+                num_classes = int(module_def["classes"])
+                img_size = int(hyperparams["height"])
+                # Define detection layer
+                yolo_layer = YOLOLayer(anchors, num_classes, img_size)
+                modules.add_module(f"yolo_{module_i}", yolo_layer)
+            # Register module list and number of output filters
+            module_list.append(modules)
+            output_filters.append(filters)
+
+        return hyperparams, module_list
+
+
     def load_darknet_weights(self, weights_path):
         """Parses and loads the weights stored in 'weights_path'"""
 
@@ -310,8 +314,14 @@ class Darknet(nn.Module):
 
         # Establish cutoff for loading backbone weights
         cutoff = None
-        if "darknet53.conv.74" in weights_path:
-            cutoff = 75
+        # If the weights file has a cutoff, we can find out about it by looking at the filename
+        # examples: darknet53.conv.74 -> cutoff is 74
+        filename = os.path.basename(weights_path)
+        if ".conv." in filename:
+            try:
+                cutoff = int(filename.split(".")[-1])+1  # use last part of filename
+            except ValueError:
+                pass
 
         ptr = 0
         for i, (module_def, module) in enumerate(zip(self.module_defs, self.module_list)):
@@ -324,30 +334,36 @@ class Darknet(nn.Module):
                     bn_layer = module[1]
                     num_b = bn_layer.bias.numel()  # Number of biases
                     # Bias
-                    bn_b = torch.from_numpy(weights[ptr : ptr + num_b]).view_as(bn_layer.bias)
+                    bn_b = torch.from_numpy(
+                        weights[ptr : ptr + num_b]).view_as(bn_layer.bias)
                     bn_layer.bias.data.copy_(bn_b)
                     ptr += num_b
                     # Weight
-                    bn_w = torch.from_numpy(weights[ptr : ptr + num_b]).view_as(bn_layer.weight)
+                    bn_w = torch.from_numpy(
+                        weights[ptr : ptr + num_b]).view_as(bn_layer.weight)
                     bn_layer.weight.data.copy_(bn_w)
                     ptr += num_b
                     # Running Mean
-                    bn_rm = torch.from_numpy(weights[ptr : ptr + num_b]).view_as(bn_layer.running_mean)
+                    bn_rm = torch.from_numpy(
+                        weights[ptr : ptr + num_b]).view_as(bn_layer.running_mean)
                     bn_layer.running_mean.data.copy_(bn_rm)
                     ptr += num_b
                     # Running Var
-                    bn_rv = torch.from_numpy(weights[ptr : ptr + num_b]).view_as(bn_layer.running_var)
+                    bn_rv = torch.from_numpy(
+                        weights[ptr : ptr + num_b]).view_as(bn_layer.running_var)
                     bn_layer.running_var.data.copy_(bn_rv)
                     ptr += num_b
                 else:
                     # Load conv. bias
                     num_b = conv_layer.bias.numel()
-                    conv_b = torch.from_numpy(weights[ptr : ptr + num_b]).view_as(conv_layer.bias)
+                    conv_b = torch.from_numpy(
+                        weights[ptr : ptr + num_b]).view_as(conv_layer.bias)
                     conv_layer.bias.data.copy_(conv_b)
                     ptr += num_b
                 # Load conv. weights
                 num_w = conv_layer.weight.numel()
-                conv_w = torch.from_numpy(weights[ptr : ptr + num_w]).view_as(conv_layer.weight)
+                conv_w = torch.from_numpy(
+                    weights[ptr : ptr + num_w]).view_as(conv_layer.weight)
                 conv_layer.weight.data.copy_(conv_w)
                 ptr += num_w
